src/main/java/com/bitdecay/checkers/rule/PiecesCannotOccupyBlackSquaresRule.java
package com.bitdecay.checkers.rule;

import com.bitdecay.board.Action;
import com.bitdecay.board.Rule;
import com.bitdecay.checkers.model.Checkerboard;

public class PiecesCannotOccupyBlackSquaresRule implements Rule<Checkerboard> {
    @Override
    public boolean apply(Checkerboard current, Checkerboard next, Action action) {
        return !next.stream().filter(piece -> (piece.x + piece.y) % 2 != 0).findFirst().isPresent();
    }

    @Override
    public String description() {
        return null;
    }

    @Override
    public String serialize() {
        return null;
    }

    @Override
    public Rule deserialize(String s) {
        return null;
    }
}
